// src/gql/design-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  SafeDate: { input: any; output: any; }
};
// ...
export type ActionPayload = Payload & {
  __typename?: 'ActionPayload';
  /**
   *
   *   An action is a ? TBD ? at a specific date.
   *   Actions are caused by a human(manual) or system(automated).
   *   Something or someone causing actions is referred to as an agent (https://english.stackexchange.com/a/96105).
   *   Action only exist in a context of a step.
   *   By humans we mean patients or care teams.
   *   By system we mean the Awell engine.
   *   Typical actions done by humans are filing in forms and checklists.
   *   Typical actions done by systems are calculations and sending messages.
   *
   *   An action definition is the set of information needed to perform an action.
   *   information can be:
   *   - static
   *   - dynamic
   *
   *   Information is static when it is not expected to change during the time a pathway is active
   *   Examples of static information:
   *   - constants. e.g., specific email "foo@gmail.com" in a send message action
   *   - gender of the patient
   *   - name of the organisation
   *   - preferences of the organisation like preferred reminder channel
   *
   *   Information is dynamic when it is collected during an execution:
   *   Examples of dynamic information:
   *   - Form submissions
   *   - step completion and activation dates
   *   - Calculation results
   *
   *
   */
  action: Action;
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type Activity = {
  __typename?: 'Activity';
  action: ActivityAction;
  action_component?: Maybe<ActionComponent>;
  container_name?: Maybe<Scalars['String']['output']>;
  context?: Maybe<PathwayContext>;
  date: Scalars['String']['output'];
  /** Url for icon, only used by extensions custom actions */
  icon_url?: Maybe<Scalars['String']['output']>;
  id: Scalars['ID']['output'];
  indirect_object?: Maybe<ActivityObject>;
  isUserActivity: Scalars['Boolean']['output'];
  label?: Maybe<ActivityLabel>;
  object: ActivityObject;
  public?: Maybe<Scalars['Boolean']['output']>;
  reference_id: Scalars['String']['output'];
  resolution?: Maybe<ActivityResolution>;
  session_id?: Maybe<Scalars['String']['output']>;
  stakeholders?: Maybe<Array<ActivityObject>>;
  status: ActivityStatus;
  stream_id: Scalars['String']['output'];
  sub_activities: Array<SubActivity>;
  subject: ActivitySubject;
  track?: Maybe<ActivityTrack>;
};
// ...
export type AhpLinkPayload = Payload & {
  __typename?: 'AhpLinkPayload';
  ahp_link: AhpLink;
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};

export type AhpLinkPayloads = Payload & {
  __typename?: 'AhpLinkPayloads';
  ahp_links: Array<AhpLink>;
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type BaselineInfoPayload = Payload & {
  __typename?: 'BaselineInfoPayload';
  baselineDataPoints: Array<BaselineDataPoint>;
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type CalculationResultPayload = Payload & {
  __typename?: 'CalculationResultPayload';
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  result: Array<CalculationSingleResult>;
  success: Scalars['Boolean']['output'];
};

export type CalculationSingleResult = {
  __typename?: 'CalculationSingleResult';
  status?: Maybe<Scalars['String']['output']>;
  subresult_id: Scalars['String']['output'];
  unit?: Maybe<Scalars['String']['output']>;
  value: Scalars['String']['output'];
  value_type?: Maybe<DataPointValueType>;
};

export type CalculationsPayload = Payload & {
  __typename?: 'CalculationsPayload';
  calculations: Array<Calculation>;
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};

export type CareflowComponentsPayload = Payload & {
  __typename?: 'CareflowComponentsPayload';
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  json: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};
// ...
export type ClinicalNotePayload = Payload & {
  __typename?: 'ClinicalNotePayload';
  clinical_note: ClinicalNote;
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type ConstantPayload = Payload & {
  __typename?: 'ConstantPayload';
  code: Scalars['String']['output'];
  constant?: Maybe<Constant>;
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};

export type ConstantsPayload = Payload & {
  __typename?: 'ConstantsPayload';
  code: Scalars['String']['output'];
  constants: Array<Constant>;
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type CreatePathwayCasePayload = Payload & {
  __typename?: 'CreatePathwayCasePayload';
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  pathway_case: PathwayCase;
  success: Scalars['Boolean']['output'];
};
// ...
export type DataCatalogJsonPayload = Payload & {
  __typename?: 'DataCatalogJsonPayload';
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  json: Scalars['String']['output'];
  success: Scalars['Boolean']['output'];
};

export type DataCatalogPayload = Payload & {
  __typename?: 'DataCatalogPayload';
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  items: Array<DataCatalogItem>;
  success: Scalars['Boolean']['output'];
};
// ...
export type DataPointDefinitionPayload = Payload & {
  __typename?: 'DataPointDefinitionPayload';
  code: Scalars['String']['output'];
  data_point_definition: DataPointDefinition;
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};

export type DataPointDefinitionsInput = {
  categories?: InputMaybe<Array<DataPointSourceType>>;
  pathway_definition_id: Scalars['String']['input'];
  valueTypes?: InputMaybe<Array<DataPointValueType>>;
};

export type DataPointDefinitionsPayload = Payload & {
  __typename?: 'DataPointDefinitionsPayload';
  code: Scalars['String']['output'];
  data_point_definitions: Array<DataPointDefinition>;
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type DynamicVariablePayload = Payload & {
  __typename?: 'DynamicVariablePayload';
  code: Scalars['String']['output'];
  dynamic_variable: DynamicVariable;
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};

export type DynamicVariablesPayload = Payload & {
  __typename?: 'DynamicVariablesPayload';
  code: Scalars['String']['output'];
  dynamic_variables: Array<DynamicVariable>;
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type ErrorObject = {
  __typename?: 'ErrorObject';
  code: Scalars['String']['output'];
  message: Scalars['String']['output'];
};
// ...
export type Mutation = {
// ...
  createPathwayCase: CreatePathwayCasePayload;
// ...
};
// ...
export type PathwayCase = {
  __typename?: 'PathwayCase';
  activities: Array<Activity>;
  created_by: PathwayCollaborator;
  id: Scalars['ID']['output'];
  last_modification_date?: Maybe<Scalars['SafeDate']['output']>;
  last_modified_by?: Maybe<PathwayCollaborator>;
  pathway_id: Scalars['String']['output'];
  start_date?: Maybe<Scalars['SafeDate']['output']>;
  status: PathwayCaseStatus;
  swimlanes: Swimlanes;
  title: Scalars['String']['output'];
};
// ...
export enum PathwayCaseStatus {
  Active = 'ACTIVE',
  Blank = 'BLANK',
  Completed = 'COMPLETED',
  MissingBaselineDatapoints = 'MISSING_BASELINE_DATAPOINTS',
  WaitingForManualTrigger = 'WAITING_FOR_MANUAL_TRIGGER'
}
// ...
export type PathwayCasesPayload = Payload & {
  __typename?: 'PathwayCasesPayload';
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  pathway_cases: Array<PathwayCase>;
  success: Scalars['Boolean']['output'];
};

export type PathwayCollaborator = {
  __typename?: 'PathwayCollaborator';
  user_email?: Maybe<Scalars['String']['output']>;
  user_id: Scalars['String']['output'];
};
// ...
export type Payload = {
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
};
// ...
export type Query = {
  __typename?: 'Query';
  action: ActionPayload;
  adHocTracks: TracksPayload;
  ahp_link: AhpLinkPayload;
  ahp_links: AhpLinkPayloads;
  baseline_info: BaselineInfoPayload;
  calculation_results: CalculationResultPayload;
  calculations: CalculationsPayload;
  careflow_components: CareflowComponentsPayload;
  clinical_note: ClinicalNotePayload;
  constant: ConstantPayload;
  constants: ConstantsPayload;
  data_catalog: DataCatalogPayload;
  /** Export data point definitions in JSON format */
  data_catalog_json: DataCatalogJsonPayload;
  data_point_definition: DataPointDefinitionPayload;
  data_point_definitions: DataPointDefinitionsPayload;
  dynamic_variable: DynamicVariablePayload;
  dynamic_variables: DynamicVariablesPayload;
  emr_report: EmrReportPayload;
  evaluatedRule: EvaluatedRulePayload;
  events: DomainEventsPayload;
  extensionActivityRecord: ExtensionActivityRecordPayload;
  extension_configurations: ExtensionConfigurationsPayload;
  extensions: ExtensionsPayload;
  form: FormPayload;
  form_data_point_definitions: DataPointDefinitionsPayload;
  form_response: FormResponsePayload;
  forms: FormsPayload;
  generated_clinical_note: GeneratedClinicalNotePayload;
  generated_emr_report: GeneratedEmrReportPayload;
  generated_message: GeneratedMessagePayload;
  invitation: InvitationPayload;
  isPathwayReadyForPreview: PathwayValidationStatusPayload;
  is_slug_unique: IsSlugUniquePayload;
  jsonPathFromMockObject: JsonPathPayload;
  jsonSchema: SchemaPayload;
  listForms: FormsPayload;
  listPathways: PathwaysPayload;
  locations: LocationsPayload;
  message: MessagePayload;
  mockObjectFromJsonSchema: MockObjectPayload;
  pathway: PathwayPayload;
  pathwayCaseActivities: ActivitiesPayload;
  pathwayCaseApiCall: PathwayCaseApiCallPayload;
  pathwayCaseApiCalls: PathwayCaseApiCallsPayload;
  pathwayCaseElements: ElementsPayload;
  pathwayCaseStepActivities: ActivitiesPayload;
  pathwayCaseTriggerActivation: TriggerActivationPayload;
  pathwayCaseTriggerActivations: TriggerActivationsPayload;
  pathwayCaseWebhookCall: PathwayCaseWebhookCallPayload;
  pathwayCaseWebhookCalls: PathwayCaseWebhookCallsPayload;
  pathwayDiffTree: PathwayDiffPayload;
  pathwayDiffs: PathwayDiffPayload;
  pathwayExtensions: ExtensionsPayload;
  pathwaySettings: PathwaySettingsPayload;
  pathwaySteps: StepsPayload;
  pathway_case: PathwayCasePayload;
  pathway_cases: PathwayCasesPayload;
  pathway_version: PathwayVersionPayload;
  pathway_versions: PathwayVersionsPayload;
  pathways: PathwaysPayload;
  rule: RulePayload;
  stakeholders: StakeholdersPayload;
  step: StepPayload;
  stepLibrary: StepLibraryPayload;
  stepLibraryForPathway: StepLibraryPayload;
  steps: StepsPayload;
  stickyNote: StickyNotePayload;
  stickyNotes: StickyNotesPayload;
  tenant: TenantPayload;
  tenants: TenantsPayload;
  timing: TimingPayload;
  track: TrackPayload;
  transition: TransitionPayload;
  transitions: TransitionsPayload;
  user: UserPayload;
  webhook: WebhookPayload;
};
// ...
export type Swimlanes = {
  __typename?: 'Swimlanes';
  items: Array<SwimlaneItem>;
  lanes: Array<Swimlane>;
  links: Array<SwimlaneLink>;
};
// ...
export type TracksPayload = Payload & {
  __typename?: 'TracksPayload';
  code: Scalars['String']['output'];
  error?: Maybe<ErrorObject>;
  success: Scalars['Boolean']['output'];
  tracks: Array<Track>;
};
// ...
export type CreatePathwayCaseMutation = { __typename?: 'Mutation', createPathwayCase: { __typename?: 'CreatePathwayCasePayload', code: string, success: boolean, pathway_case: { __typename?: 'PathwayCase', id: string, pathway_id: string, title: string, last_modification_date?: any | null, start_date?: any | null, status: PathwayCaseStatus } } };
// ...
export type PathwayCasesQuery = { __typename?: 'Query', pathway_cases: { __typename?: 'PathwayCasesPayload', code: string, success: boolean, pathway_cases: Array<{ __typename?: 'PathwayCase', id: string, pathway_id: string, title: string, last_modification_date?: any | null, start_date?: any | null, status: PathwayCaseStatus }> } };